# feat: split E2StoreStream into reader and writer