### What was wrong?

The `E2StoreStream` uses `File` structure for reading/writing into file. This can be more optimized by using `BufReader`/`BufWriter`.

### How was it fixed?

I split `E2StoreStream` into `E2StoreStreamReader` and `E2StoreStreamWriter` that use `BufReader`/`BufWriter`. I also had to split `Era2` into `Era2Reader` and `Era2Writer`.

In my opinion, this split makes more sense as it would most likely be wrong to read and write into the same file at the same time.

To measure if this actually improves performance as I expect, I did the following:

0. generated era2 files for blocks 3 million, 4 million and 5 million
1. built release version of `trin-execution` binary at both `master` and this `pr`
2. for both builds (master/pr), imported each era2 state into clean directory
3. exported era2 from directory from step2

Here are the results

| block number | import master | import pr | export master | export pr |
|--------------|---------------|-----------|---------------|-----------|
| 3'000'000    | 2m11s         | 2m10s     | 37s           | 18s       |
| 4'000'000    | 8m55s         | 9m01s     | 2m44s         | 1m35s     |
| 5'000'000    | 54m22s        | 52m55s    | 43m38s        | 37m02s    |

We can see that there is no big difference when in comes to importing era2 file (reading), but there is noticeable improvement for exporting (writing). 

### To-Do

[//]: # (Stay ahead of things, add list items here!)
- [ ] Clean up commit history and use [conventional commits](https://www.conventionalcommits.org/en/v1.0.0/).