@KolbyML 2024-10-06
If there is no difference in performance using `BufReader` that should be removed from the PR