@KolbyML 2024-10-06
How big are the .era2 files for 3,4,5 million can you upload them to era2.ethportal.net.

Will there be a performance difference at block 10,15,20 million well exporting?