@morph-dev 2024-10-07
> If there is no difference in performance using `BufReader` that should be removed from the PR

I would say that it probably has improvements, but they might not be significant. And in my setup, what took the most time was recreating the storage trie and writing it to rocks db.

Sure, since it might not be as relevant, I can remove `BufReader`. But I would still keep `E2StoreStreamReader` and `E2StoreStreamWriter` separate (as they need to be because of the `BufWriter`). In which case, we might just keep `BufReader` because it should be better than using `File`.


> How big are the .era2 files for 3,4,5 million can you upload them to era2.ethportal.net.

3 million - 358 MB / 4 million - 1.5 GB / 5 million - 7.5 GB

> Will there be a performance difference at block 10,15,20 million well exporting?

I didn't try as it would take too long, but I assume there will be.
To have good results of my benchmark, I would have to not use my computer during that period (as there is a lot of CPU/RAM usage while processing trie), and me using the computer can affect it.

I guess I can do a test in which I'm just reading/writing to RAM memory (not actually processing the storage) and remove the noise. I can look into it if you think it's necessary.