@morph-dev 2024-10-07
> How big are the .era2 files for 3,4,5 million can you upload them to era2.ethportal.net.

And yes, I plan to upload them, once I export few more.