@morph-dev 2024-10-07
I created a binary that just reads the era2 file and collects stats on how many account entries, storage entries, storage items and non-empty bytecodes there are.
Exported commit [here](https://github.com/morph-dev/trin/commit/e3c0c057bf47d0efbb75ab01ae84ef7c3c954a0b), and I think it can be useful and I plan to make a pr, unless you have some objection.

This is the cleanest benchmark for `File` vs `BufReader` in context of era2 file that I can think of. Here are the result:
| block_number | master | pr     |
|--------------|--------|--------|
| 3'000'000    | 18s    | 17s    |
| 4'000'000    | 1m15s  | 1m12s  |
| 5'000'000    | 5m15s  | 5m0s   |
| 10'000'000   | 21m2s  | 19m51s |

The benefits of `BufReader` are noticeable, but they are not huge.

According to [documentation](https://doc.rust-lang.org/std/io/struct.BufReader.html): `BufReader<R> can improve the speed of programs that make small and repeated read calls to the same file or network socket`. I think this fits nicely in our use case, expect that we have a mix of small (account_entry) and big reads (storage_entry). 
If we decide to support Stream reader on some type that is not era2, we might have bigger improvements. Or if we decide to stream it directly from the internet.

No matter how I look at it, it is or can be beneficial to use `BufReader`, without downside (except that it requires extra line of code for initialization).