@KolbyML 2024-10-07
> I created a binary that just reads the era2 file and collects stats on how many account entries, storage entries, storage items and non-empty bytecodes there are. Exported commit [here](https://github.com/morph-dev/trin/commit/e3c0c057bf47d0efbb75ab01ae84ef7c3c954a0b), and I think it can be useful and I plan to make a pr, unless you have some objection.

You can make a PR for it if you want, I don't see a use case for it right now but that is me.

> This is the cleanest benchmark for `File` vs `BufReader` in context of era2 file that I can think of. Here are the result:
> 
> block_number	master	pr
> 3'000'000	18s	17s
> 4'000'000	1m15s	1m12s
> 5'000'000	5m15s	5m0s
> 10'000'000	21m2s	19m51s
> The benefits of `BufReader` are noticeable, but they are not huge.

I think these numbers are within the margin of error i.e. it doesn't make a difference

> According to [documentation](https://doc.rust-lang.org/std/io/struct.BufReader.html): `BufReader<R> can improve the speed of programs that make small and repeated read calls to the same file or network socket`. I think this fits nicely in our use case, expect that we have a mix of small (account_entry) and big reads (storage_entry). 

In testing there is no difference within the margin of error

> If we decide to support Stream reader on some type that is not era2, we might have bigger improvements. 

I don't see how that is related to this PR

> Or if we decide to stream it directly from the internet.

Is there a benchmark for this?

> No matter how I look at it, it is or can be beneficial to use `BufReader`, without downside (except that it requires extra line of code for initialization).

If there was no downside to `BufReader` it would be the default functionality of `File`. This PR is an optimization PR so if a performance increase can't be measured (10 seconds on 10 million isn't a performance increase) then it doesn't make sense to add the complexity.


If you want to keep it go for it, but fundamentally it doesn't make sense to keep it unless there is something measurable. Maybe testing it on a harddrive would help for a worse case possibility it could help?