@KolbyML 2024-10-07
Why is importing at 5 million takes over 50 minutes, but 10 million takes over 20 minutes?